chore/#196 : 리스트 삭제 mutation hook으로 분리
