+++ utils/hooks/queries/inventory/inventory.ts
@@ -1,4 +1,6 @@
-import { useQuery } from 'react-query';
+import { useMutation, useQuery, useQueryClient } from 'react-query';
+import { DeleteAloneInventoryInput } from '../../../../service/inventory/alone';
+import { DeleteTogetherInventoryInput } from '../../../../service/inventory/together';
 import useAPI from '../../useAPI';
 
 export const useGetTogetherInventory = (id: string) => {
@@ -17,4 +19,46 @@ export const useGetAloneInventory = (id: string) => {
   });
 
   return data;
+};
+
+export const useDeleteTogetherInventory = (payload: { folderId: string; listId: string }) => {
+  const queryClient = useQueryClient();
+
+  const deleteTogetherInventory = useAPI(
+    (api) => (params: DeleteTogetherInventoryInput) =>
+      api.inventory.together.deleteTogetherInventory(params),
+  );
+  const { mutate: deleteTogetherInventoryMutate } = useMutation(
+    () => {
+      return deleteTogetherInventory(payload);
+    },
+    {
+      onSuccess: () => {
+        queryClient.invalidateQueries('getTogetherInventory');
+      },
+    },
+  );
+
+  return deleteTogetherInventoryMutate;
+};
+
+export const useDeleteAloneInventory = (payload: { folderId: string; listId: string }) => {
+  const queryClient = useQueryClient();
+
+  const deleteAloneInventory = useAPI(
+    (api) => (params: DeleteAloneInventoryInput) =>
+      api.inventory.alone.deleteAloneInventory(params),
+  );
+  const { mutate: deleteAloneInventoryMutate } = useMutation(
+    () => {
+      return deleteAloneInventory(payload);
+    },
+    {
+      onSuccess: () => {
+        queryClient.invalidateQueries('getAloneInventory');
+      },
+    },
+  );
+
+  return deleteAloneInventoryMutate;
 };
